python/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
g4f
diskcache
anthropic
numpy==1.24.2 
packaging==23.1 
pandas==1.5.3 
psycopg2-binary==2.9.6 
platformdirs==2.6.0 
plac==1.3.5 
pipenv==2022.11.30 
parsel==1.8.1 
Protego==0.2.1 
prompt-toolkit==3.0.38 
nvidia-cublas-cu11==11.10.3.66 
nvidia-cuda-nvrtc-cu11==11.7.99 
nvidia-cuda-runtime-cu11==11.7.99 
nvidia-cudnn-cu11==8.5.0.96
python-multipart